fix the backlight off feature, renaming the apps to 3dsMouse and little cleaning
the originol author is still credited

+++ PC/include/settings.h
@@ -10,18 +10,18 @@ struct settings {
 };
 
 struct StartCoor {
-    double x;
-    double y;
+    int x;
+    int y;
 };
 
 struct EndCoor {
-    double x;
-    double y;
+    int x;
+    int y;
 };
 
 struct touch {
-	short x;
-	short y;
+	unsigned short x;
+	unsigned short y;
 };
 
 extern struct touch currentTouch;
